feat: scan qr-codes on send screen
